requirements.txt
fastapi
uvicorn
python-dotenv
opencv-python
insightface
numpy
requests
